add separate nodedefaults file
