+++ __init__.py
@@ -104,8 +104,8 @@ ui_modules = [
     # bgl modules
     "viewer_draw", "viewer_draw_mk2", "nodeview_bgl_viewer_draw", "nodeview_bgl_viewer_draw_mk2",
     "index_viewer_draw", "bgl_callback_3dview",
-    # show git info
-    "development",
+    # show git info, general node details
+    "development", "node_defaults"
 ]
 
 # modules and pkg path, nodes are done separately.
